Signedness and integer type conversion warning fixes for VC++
git-svn-id: svn://svn.gna.org/svn/mysqlpp/trunk@855 5946f024-35f4-0310-b1b3-9a3be3380cfb

+++ lib/manip.cpp
@@ -63,7 +63,8 @@ SQLQueryParms& operator <<(quote_type2 p, SQLString& in)
 		}
 		else {
 			char* s = new char[in.size() * 2 + 1];
-			mysql_escape_string(s, in.c_str(), in.size());
+			mysql_escape_string(s, in.c_str(),
+					(unsigned long)in.size());
 			SQLString in2 = SQLString('\'') + s + '\'';
 			in2.processed = true;
 			*p.qparms << in2;
@@ -103,9 +104,9 @@ ostream& operator <<(quote_type1 o, const string& in)
 template <>
 ostream& operator <<(quote_type1 o, const char* const& in)
 {
-	unsigned int size = strlen(in);
+	size_t size = strlen(in);
 	char* s = new char[size * 2 + 1];
-	mysql_escape_string(s, in, size);
+	mysql_escape_string(s, in, (unsigned long)size);
 	*o.ostr << '\'' << s << '\'';
 	delete[] s;
 	return *o.ostr;
@@ -119,7 +120,7 @@ inline ostream& _manip(quote_type1 o, const ColData_Tmpl<Str>& in)
 {
 	if (in.escape_q()) {
 		char* s = new char[in.size() * 2 + 1];
-		mysql_escape_string(s, in.c_str(), in.size());
+		mysql_escape_string(s, in.c_str(), (unsigned long)in.size());
 		if (in.quote_q())
 			*o.ostr << '\'' << s << '\'';
 		else
@@ -179,7 +180,7 @@ ostream& operator <<(ostream& o, const ColData_Tmpl<string>& in)
 
 	if (in.escape_q()) {
 		char* s = new char[in.size() * 2 + 1];
-		mysql_escape_string(s, in.c_str(), in.size());
+		mysql_escape_string(s, in.c_str(), (unsigned long)in.size());
 		if (in.quote_q())
 			o << '\'' << s << '\'';
 		else
@@ -243,7 +244,7 @@ Query& operator <<(Query& o, const ColData_Tmpl<string>& in)
 	}
 	if (in.escape_q()) {
 		char* s = new char[in.size() * 2 + 1];
-		mysql_escape_string(s, in.c_str(), in.size());
+		mysql_escape_string(s, in.c_str(), (unsigned long)in.size());
 		if (in.quote_q())
 			static_cast<ostream&>(o) << '\'' << s << '\'';
 		else
@@ -413,7 +414,7 @@ SQLQueryParms& operator <<(escape_type2 p, SQLString& in)
 {
 	if (in.is_string && ! in.dont_escape) {
 		char* s = new char[in.size() * 2 + 1];
-		mysql_escape_string(s, in.c_str(), in.size());
+		mysql_escape_string(s, in.c_str(), (unsigned long)in.size());
 		SQLString in2 = s;
 		in2.processed = true;
 		*p.qparms << in2;
@@ -455,9 +456,9 @@ std::ostream& operator <<(escape_type1 o, const std::string& in)
 template <>
 ostream& operator <<(escape_type1 o, const char* const& in)
 {
-	unsigned int size = strlen(in);
+	size_t size = strlen(in);
 	char* s = new char[size * 2 + 1];
-	mysql_escape_string(s, in, size);
+	mysql_escape_string(s, in, (unsigned long)size);
 	*o.ostr << s;
 	delete[] s;
 	return *o.ostr;
@@ -471,7 +472,7 @@ inline ostream& _manip(escape_type1 o, const ColData_Tmpl<Str>& in)
 {
 	if (in.escape_q()) {
 		char* s = new char[in.size() * 2 + 1];
-		mysql_escape_string(s, in.c_str(), in.size());
+		mysql_escape_string(s, in.c_str(), (unsigned long)in.size());
 		delete[] s;
 	}
 	else {
